requirements.txt
customtkinter>=5.2.0
PyYAML>=6.0
